refactor: UserService interface 활용하도록 변경
