+++ app/src/test/java/com/techcourse/service/UserServiceTest.java
@@ -38,7 +38,9 @@ class UserServiceTest {
     void testChangePassword() {
         final var userHistoryDao = new UserHistoryDao(jdbcTemplate);
         userHistoryDao.log(new UserHistory(1L, 1L, "aa", "bb", "cc", "dd"));
-        final var userService = new UserService(userDao, userHistoryDao, transactionalExecutor);
+        final var appUserService = new AppUserService(userDao, userHistoryDao);
+
+        final var userService = new TxUserService(appUserService, transactionalExecutor);
 
         final var newPassword = "qqqqq";
         final var createBy = "gugu";
@@ -53,13 +55,16 @@ class UserServiceTest {
     void testTransactionRollback() {
         // 트랜잭션 롤백 테스트를 위해 mock으로 교체
         final var userHistoryDao = new MockUserHistoryDao(jdbcTemplate);
-        final var userService = new UserService(userDao, userHistoryDao, transactionalExecutor);
+        // 애플리케이션 서비스
+        final var appUserService = new AppUserService(userDao, userHistoryDao);
+        // 트랜잭션 서비스 추상화
+        final var userService = new TxUserService(appUserService, transactionalExecutor);
 
         final var newPassword = "newPassword";
         final var createBy = "gugu";
         // 트랜잭션이 정상 동작하는지 확인하기 위해 의도적으로 MockUserHistoryDao에서 예외를 발생시킨다.
         assertThrows(DataAccessException.class,
-                () -> userService.changePassword(1L, newPassword, createBy));
+            () -> userService.changePassword(1L, newPassword, createBy));
 
         final var actual = userService.findById(1L);
